requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
groq
httpx
python-dotenv
pydantic
pydantic-settings
pandas
openpyxl
